requirements.txt
asgiref==3.8.1
boto3==1.37.37
botocore==1.37.37
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
Django==5.2
dotenv==0.9.9
idna==3.10
jmespath==1.0.1
pillow==11.2.1
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
reportlab==4.4.0
requests==2.32.3
s3transfer==0.11.5
setuptools==79.0.0
six==1.17.0
sqlparse==0.5.3
urllib3==2.4.0
wheel==0.45.1
